xbee api
git-svn-id: svn://localhost:5050/magic2010/trunk@588 61389559-6c2a-4883-acdc-2d27d9f19ada

+++ drivers/Xbee/testXbeeApi.m
@@ -20,7 +20,7 @@ while(1)
   w = w/4;
   if (ROBOT_CNTRL.selected > 0)
     if (prevSelected ~= ROBOT_CNTRL.selected)
-      XbeeAPI('writeRobSelect',uint8(ROBOT_CNTRL.selected));
+      XbeeAPI('writeRobSelect',ROBOT_CNTRL.selected);
       prevSelected = ROBOT_CNTRL.selected;
     end
